third_party/WebKit/LayoutTests/flag-specific/enable-blink-features=LayoutNG/fast/events/shift-drag-selection-on-link-triggers-drag-n-drop-expected.txt

This is a testharness.js-based test.
FAIL Untitled assert_equals: expected 1 but got 0
Harness: the test ran to completion.
